pyproject.toml
[project]
name = "hexlet-vacancy-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asgiref>=3.8.1",
    "bs4>=0.0.2",
    "argon2-cffi>=25.1.0",
    "dj-database-url>=3.0.0",
    "django>=5.2.1",
    "django-filter>=25.1",
    "dotenv>=0.9.9",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.3",
    "telethon>=1.40.0",
    "django-ratelimit>=4.1.0",
    "inertia-django>=1.2.0",
    "django-vite>=3.1.0",
    "whitenoise>=6.11.0",
    "gunicorn>=23.0.0",
    "mypy>=1.19.1",
    "django-stubs>=5.2.8",
    "types-requests>=2.32.4.20260107",
    "pre-commit>=4.3.0",
    "aiohttp>=3.13.2",
    "celery>=5.5.3",
    "redis>=6.4.0",
    "celery-stubs>=0.1.3",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "pytest-django>=4.11.1",
    "ruff>=0.12.1",
]

[tool.pylint.main]
  source-roots = "app/services"
